pyproject.toml
[tool.poetry]
name = "griptape-rag-example-pgvector"
version = "0.1.0"
description = "A short sample showing how to use Griptape with pgvector"
authors = ["Ian Massingham"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
griptape = {extras = ["all"], version = "^1.2.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
